Add sorting function and associated tests
A sortAscending function was implemented to sort a 2D string slice in ascending order based on values in a specific column. This function is then used in the encoder to sort the TGU array by callsign before writing it to a file. Accompanying tests were also added to check the functionality of the new sort function.